refactor(auth): rename files and add ssl in db config
