+++ src/modules/auth/auth.module.ts
@@ -2,10 +2,10 @@ import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UserEntity } from '../users/users.entity';
+import { UserEntity } from '../account/entity/user.entity';
 import { AuthRepository } from './auth.repository';
 import { MailModule } from '../mail/mail.module';
-import { RoleEntity } from '../role/roles.entity';
+import { RoleEntity } from '../role/role.entity';
 import { RefreshTokenEntity } from '../token/refresh-token.entity';
 import { GoogleAuthService } from './oauth/google.service';
 import { OAuthAccountEntity } from './oauth/oauth.entity';
