+++ src/modules/auth/dto/auth.signout.dto.ts
