+++ src/modules/config/db.config.ts
@@ -6,6 +6,7 @@ export const typeOrmConfig = (
 ): TypeOrmModuleOptions => ({
   type: 'postgres',
   host: configService.get<string>('DB_HOST'),
+  ssl: configService.get<boolean>('DB_SSL'),
   port: Number(configService.get<string>('DB_PORT')),
   username: configService.get<string>('DB_USER'),
   password: configService.get<string>('DB_PASS'),
