+++ src/modules/role/role.guard.ts
@@ -5,8 +5,8 @@ import {
   ForbiddenException,
 } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
-import { Role } from './roles.enum';
-import { ROLES_KEY } from './roles.decorator';
+import { Role } from './role.enum';
+import { ROLES_KEY } from './role.decorator';
 import { Request } from 'express';
 import { AuthUser } from '../token/authuser.interface';
 import { message } from 'src/modules/common/constants.common';
